requirements.txt
fastapi
langchain
PyPDF2
aiofiles
axios
python-dotenv